Issue 24
This commit adjust an issue that was preventing proxysql-admin to run
on version 1.4.14.
It also adds a small delay after the proxysql service has been started.
During my tests, sometimes proxysql has not been able to start the
chield process and we were already calling proxysql-admin, which
was resulting in an error to connect to proxysql.